+++ .gitignore
@@ -2,6 +2,7 @@
 build/
 /.idea
 Mango.iml
+out/
 
 # Ignore Gradle GUI config
 gradle-app.setting
