+++ build.gradle
@@ -23,5 +23,6 @@ dependencies {
     compile group: 'commons-io', name: 'commons-io', version: '2.4'
     compile group: 'org.freemarker', name:'freemarker', version: '2.3.14'
     compile group: 'mysql', name: 'mysql-connector-java', version: '5.1.44'
+    compile group: 'org.jasypt', name: 'jasypt', version: '1.9.2'
 
 }
