+++ src/main/java/com/tgp/controller/UserController.java
@@ -10,36 +10,36 @@ import java.util.List;
 
 public class UserController {
 
-   private static UserDao userDao = Dbi.getInstance().open(UserDao.class);
+    private static UserDao userDao = Dbi.getInstance().open(UserDao.class);
 
-   public List<Pair<Integer, String>> getListOfNames() {
-      List<User> users = userDao.getAllUsers();
+    public List<Pair<Integer, String>> getListOfNames() {
+        List<User> users = userDao.getAllUsers();
 
-      List<Pair<Integer, String>> names = new ArrayList<>();
-      for (User user : users) {
-         names.add(new Pair(user.getId(),user.getLastName() + ", " + user.getFirstName()));
-      }
+        List<Pair<Integer, String>> names = new ArrayList<>();
+        for (User user : users) {
+            names.add(new Pair(user.getId(), user.getLastName() + ", " + user.getFirstName()));
+        }
 
-      return names;
-   }
+        return names;
+    }
 
-   public int insertUser(String firstName, String lastName) {
-      return userDao.insertUser(firstName, lastName);
-   }
+    public int insertUser(String firstName, String lastName) {
+        return userDao.insertUser(firstName, lastName);
+    }
 
-   public User findUserByName(String firstName, String lastName) {
-      return userDao.findUserByName(firstName, lastName);
-   }
+    public User findUserByName(String firstName, String lastName) {
+        return userDao.findUserByName(firstName, lastName);
+    }
 
-   public int deleteUser(User user) {
-      return userDao.deleteUser(user.getId());
-   }
+    public int deleteUser(User user) {
+        return userDao.deleteUser(user.getId());
+    }
 
-   public User findUserById(int id) {
-      return userDao.findUserById(id);
-   }
+    public User findUserById(int id) {
+        return userDao.findUserById(id);
+    }
 
-   public List<User> getAllUsers() {
-      return userDao.getAllUsers();
-   }
+    public List<User> getAllUsers() {
+        return userDao.getAllUsers();
+    }
 }
