+++ src/main/java/com/tgp/db/dao/BaseDao.java
@@ -2,5 +2,6 @@ package com.tgp.db.dao;
 
 public interface BaseDao {
 
+    @SuppressWarnings("unused")
     void close();
 }
