+++ src/main/java/com/tgp/db/dao/mapper/UserMapper.java
@@ -9,14 +9,14 @@ import java.sql.SQLException;
 
 public class UserMapper implements ResultSetMapper<User> {
 
-   @Override
-   public User map(int index, ResultSet r, StatementContext ctx) throws SQLException {
-      int id = r.getInt("id");
-      String firstName = r.getString("first_name");
-      String lastName = r.getString("last_name");
+    @Override
+    public User map(int index, ResultSet r, StatementContext ctx) throws SQLException {
+        int id = r.getInt("id");
+        String firstName = r.getString("first_name");
+        String lastName = r.getString("last_name");
 
-      User user = new User(firstName, lastName);
-      user.setId(id);
-      return user;
-   }
+        User user = new User(firstName, lastName);
+        user.setId(id);
+        return user;
+    }
 }
