+++ src/main/java/com/tgp/util/FtlConfig.java
@@ -8,13 +8,13 @@ import java.util.Map;
 
 public class FtlConfig {
 
-   public static String get(String templatePath, Map<String, Object> data) throws Exception {
-      Configuration configuration = new Configuration();
-      configuration.setClassForTemplateLoading(FtlConfig.class.getClass(), "/");
-      Template template = configuration.getTemplate(templatePath);
-      StringWriter output = new StringWriter();
-      template.process(data, output);
+    public static String get(String templatePath, Map<String, Object> data) throws Exception {
+        Configuration configuration = new Configuration();
+        configuration.setClassForTemplateLoading(FtlConfig.class.getClass(), "/");
+        Template template = configuration.getTemplate(templatePath);
+        StringWriter output = new StringWriter();
+        template.process(data, output);
 
-      return output.toString();
-   }
+        return output.toString();
+    }
 }
