+++ src/main/java/com/tgp/util/Pair.java
@@ -1,21 +1,21 @@
 package com.tgp.util;
 
-public class Pair<K, V>{
+public class Pair<K, V> {
 
-   private K _1;
-   private V _2;
+    private K _1;
+    private V _2;
 
-   public Pair(K _1, V _2) {
-      this._1 = _1;
-      this._2 = _2;
-   }
+    public Pair(K _1, V _2) {
+        this._1 = _1;
+        this._2 = _2;
+    }
 
-   public K _1() {
-      return _1;
-   }
+    public K _1() {
+        return _1;
+    }
 
-   public V _2() {
-      return _2;
-   }
+    public V _2() {
+        return _2;
+    }
 
 }
