+++ src/main/java/com/tgp/view/pane/AdminPane.java
@@ -8,22 +8,22 @@ import javafx.scene.layout.GridPane;
 
 public class AdminPane extends BorderPane {
 
-   public AdminPane() {
-      Button manageUsersButton = new Button("Manage Users");
+    public AdminPane() {
+        Button manageUsersButton = new Button("Manage Users");
 
-      manageUsersButton.setOnAction(e -> {
-         getScene().getWindow().hide();
-         new ManageUsersStage();
-      });
+        manageUsersButton.setOnAction(e -> {
+            getScene().getWindow().hide();
+            new ManageUsersStage();
+        });
 
-      GridPane gridPane = new GridPane();
-      gridPane.setAlignment(Pos.CENTER);
-      gridPane.setHgap(10);
-      gridPane.setVgap(12);
+        GridPane gridPane = new GridPane();
+        gridPane.setAlignment(Pos.CENTER);
+        gridPane.setHgap(10);
+        gridPane.setVgap(12);
 
-      gridPane.add(manageUsersButton, 0, 0);
+        gridPane.add(manageUsersButton, 0, 0);
 
-      setCenter(gridPane);
-   }
+        setCenter(gridPane);
+    }
 
 }
