+++ src/main/java/com/tgp/view/stage/AdminStage.java
@@ -5,10 +5,10 @@ import javafx.stage.Stage;
 
 public class AdminStage extends Stage {
 
-   public AdminStage() {
-      setTitle("Admin");
-      setScene(new AdminScene());
-      show();
-      setOnCloseRequest(f -> new MainStage());
-   }
+    public AdminStage() {
+        setTitle("Admin");
+        setScene(new AdminScene());
+        show();
+        setOnCloseRequest(f -> new MainStage());
+    }
 }
