+++ src/main/java/com/tgp/view/viewmodel/LogViewModel.java
@@ -1,0 +1,31 @@
+package com.tgp.view.viewmodel;
+
+public class LogViewModel {
+    private String logType;
+    private String time;
+    private String imagePath;
+
+    public void setImagePath(String imagePath) {
+        this.imagePath = imagePath;
+    }
+
+    public void setLogType(String logType) {
+        this.logType = logType;
+    }
+
+    public void setTime(String time) {
+        this.time = time;
+    }
+
+    public String getImagePath() {
+        return imagePath;
+    }
+
+    public String getLogType() {
+        return logType;
+    }
+
+    public String getTime() {
+        return time;
+    }
+}
