+++ src/main/resources/app.properties
@@ -1,6 +1,7 @@
-app.admin.password=admin0987
+app.admin.password=BqF8KoW2/F/GuJuBMczBnuB7OtfxnTRddAvhysYlFBUZRMhzH/CZLsyquWtJibat
 
 mysql.db.name=tgp
 mysql.db.user=root
 mysql.db.password=root1234
+mysql.db.host={host}
 mysql.db.port=3306
